zoom/requirements.txt
Flask==2.3.2
requests==2.31.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.4
python-socketio==5.9.0
python-engineio==4.7.1
python-dotenv